StepsProxy: move into separate source
